+++ app/src/main/java/com/teamwss/websoso/ui/main/myPage/myLibrary/MyLibraryViewModel.kt
@@ -1,4 +1,4 @@
-package com.teamwss.websoso.ui.myPage.myLibrary
+package com.teamwss.websoso.ui.main.myPage.myLibrary
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModel
 import com.teamwss.websoso.data.model.GenrePreferenceEntity
 import com.teamwss.websoso.data.model.NovelPreferenceEntity
 import com.teamwss.websoso.data.repository.FakeUserRepository
-import com.teamwss.websoso.ui.myPage.myLibrary.model.AttractivePoints
+import com.teamwss.websoso.ui.main.myPage.myLibrary.model.AttractivePoints
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
